UI: show winner(s) in WinnerCard.jsx
